Migrate advent command and add reminders
